+++ packages/components/src/ProForm/widgets/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Fields, FormItemsProps } from '../type';
 import {
   Input,
@@ -61,6 +60,7 @@ export function getFormFields(
       if (!hide) {
         const name = key;
         const Widget = widgetsList[widget];
+        console.log(widgetProps);
         fields[name] = {
           label: label,
           children: <Widget {...widgetProps} />,