+++ packages/components/src/ProTable/widgets/Radio.tsx
@@ -1,16 +1,33 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { Radio, RadioGroup } from 'uiw';
 
 interface FormSelectProps {
   option?: HTMLOptionElement[];
+  disabled?: boolean;
 }
 
-const FormRadio: React.FC<FormSelectProps> = ({ option, ...others }) => {
+const FormRadio: React.FC<FormSelectProps> = ({
+  disabled,
+  option,
+  ...others
+}) => {
+  const [allDisabled, setAllDisabled] = useState<boolean>(false);
+
+  // 全部枚举设为disabled
+
+  useEffect(() => {
+    setAllDisabled(disabled || false);
+  }, [disabled]);
+
   return (
     <RadioGroup {...others}>
       {option &&
         option.map((opt) => (
-          <Radio key={opt.value} value={opt.value}>
+          <Radio
+            disabled={allDisabled || opt?.disabled}
+            key={opt.value}
+            value={opt.value}
+          >
             {opt.label}
           </Radio>
         ))}